manager: extend versionCode to Long to exceed Int.MAX
Fixed an issue where versionCode would get stuck at 2147483647. Updated data type to Long to support larger version numbers.
Some modules call multiple parameters for versionCode, causing it to reach 10 digits, which may cause issues on Android 21+. Please check for any omissions.

+++ app/core/src/main/java/com/topjohnwu/magisk/core/model/module/LocalModule.kt
@@ -20,7 +20,7 @@ data class LocalModule(
     override var id: String = ""
     override var name: String = ""
     override var version: String = ""
-    override var versionCode: Int = -1
+    override var versionCode: Long = -1L
     var author: String = ""
     var description: String = ""
     var updateInfo: OnlineModule? = null
@@ -79,7 +79,7 @@ data class LocalModule(
                 "id" -> id = value
                 "name" -> name = value
                 "version" -> version = value
-                "versionCode" -> versionCode = value.toInt()
+                "versionCode" -> versionCode = value.toLong()
                 "author" -> author = value
                 "description" -> description = value
                 "updateJson" -> updateUrl = value
